feat(detail-page-template)!: replace subheader/bodyTabs with avatar/badges/meta + tabs slot
BREAKING CHANGE: `subheader` and `bodyTabs` props removed from
DetailPageTemplate. Migrate to: `avatar`/`badges`/`meta` slots on
PageHeader (via DetailPageTemplate) for what previously lived in
`subheader`, and nav-link tabs (a `<Tabs.Root>` with only `<Tabs.List>`
inside, passed to `tabs`) for what previously used `bodyTabs`.

Also moves BodyTabsItem/BodyTabsProp type definitions into
settings-page-template.tsx (the only remaining consumer).
