+++ src/templates/detail-page-template.test.tsx
@@ -1,8 +1,8 @@
 // src/templates/detail-page-template.test.tsx
 import { ChakraProvider, defaultSystem } from "@chakra-ui/react";
-import { render, screen } from "@testing-library/react";
+import { render, screen, within } from "@testing-library/react";
 import type { ReactElement } from "react";
-import { describe, expect, it, vi } from "vitest";
+import { describe, expect, it } from "vitest";
 import { AppShell } from "./app-shell";
 import { DetailPageTemplate } from "./detail-page-template";
 
@@ -53,7 +53,7 @@ describe("DetailPageTemplate", () => {
 		expect(wrapper.getAttribute("style") ?? "").not.toMatch(/padding/);
 	});
 
-	it("renders the tabs slot when provided", () => {
+	it("renders the tabs slot inside the registered header", () => {
 		renderWithChakra(
 			<AppShell sidebar={<div />}>
 				<DetailPageTemplate
@@ -64,71 +64,28 @@ describe("DetailPageTemplate", () => {
 				</DetailPageTemplate>
 			</AppShell>,
 		);
-		expect(screen.getByTestId("tabs")).toBeInTheDocument();
+		const header = screen.getByTestId("app-shell-header");
+		expect(within(header).getByTestId("tabs")).toBeInTheDocument();
 	});
 
-	it("renders the active tab's content via bodyTabs (uncontrolled)", () => {
+	it("forwards avatar, badges, meta, and tabs into the registered PageHeader", () => {
 		renderWithChakra(
 			<AppShell sidebar={<div />}>
 				<DetailPageTemplate
-					title="X"
-					bodyTabs={{
-						defaultValue: "a",
-						items: [
-							{
-								value: "a",
-								label: "Tab A",
-								content: <div data-testid="content-a">A</div>,
-							},
-							{
-								value: "b",
-								label: "Tab B",
-								content: <div data-testid="content-b">B</div>,
-							},
-						],
-					}}
-				/>
-			</AppShell>,
-		);
-		expect(screen.getByTestId("content-a")).toBeInTheDocument();
-		expect(screen.queryByTestId("content-b")).not.toBeInTheDocument();
-	});
-
-	it("renders the subheader between header and tabs", () => {
-		renderWithChakra(
-			<AppShell sidebar={<div />}>
-				<DetailPageTemplate
-					title="X"
-					subheader={<div data-testid="subheader">subheader</div>}
+					title="Jana Schmid"
+					avatar={<div data-testid="av">JS</div>}
+					badges={<span data-testid="bd">Aktiv</span>}
+					meta={<span data-testid="mt">jana@example.test</span>}
+					tabs={<div data-testid="tb">tab list</div>}
 				>
-					<div data-testid="body">body</div>
+					body
 				</DetailPageTemplate>
 			</AppShell>,
 		);
-		const sub = screen.getByTestId("subheader");
-		const body = screen.getByTestId("body");
-		expect(sub).toBeInTheDocument();
-		expect(
-			sub.compareDocumentPosition(body) & Node.DOCUMENT_POSITION_FOLLOWING,
-		).toBeTruthy();
-	});
-
-	it("throws when both bodyTabs and tabs are passed", () => {
-		const spy = vi.spyOn(console, "error").mockImplementation(() => {});
-		expect(() =>
-			renderWithChakra(
-				<AppShell sidebar={<div />}>
-					<DetailPageTemplate
-						title="X"
-						tabs={<div />}
-						bodyTabs={{
-							defaultValue: "a",
-							items: [{ value: "a", label: "A", content: <div /> }],
-						}}
-					/>
-				</AppShell>,
-			),
-		).toThrow(/mutually exclusive/i);
-		spy.mockRestore();
+		const header = screen.getByTestId("app-shell-header");
+		expect(within(header).getByTestId("av")).toBeInTheDocument();
+		expect(within(header).getByTestId("bd")).toBeInTheDocument();
+		expect(within(header).getByTestId("mt")).toBeInTheDocument();
+		expect(within(header).getByTestId("tb")).toBeInTheDocument();
 	});
 });
